update ApiClientConfigurator and part of clients
